Throughout Cygwin, use u_intN_t or uintN_t
Don't use u_char, u_short, u_int or u_long in Cygwin, unless it refers
to the Winsock types.  Use u_intN_t in BSD-based sources, unsigned char
where strings are concerned, uintN_t otherwise.  Also:

	* net.cc: Fix comment, we're not using u_long anymore.

+++ winsup/cygwin/libc/inet_addr.c
@@ -100,7 +100,7 @@ __FBSDID("$FreeBSD$");
  */
 extern int
 cygwin_inet_aton(const char *cp, struct in_addr *addr) {
-	u_long val;
+	u_int32_t val;
 	int base, n;
 	char c;
 	u_int8_t parts[4];
